First cut as providing support for building under MSVC 2019.
+ Provide a `CMakeLists.txt` with basic build instructions.
+ Clean up some warnings.
+ Provide MSVC alternatives for some features that were specific to Linux.

+++ io.c
@@ -33,7 +33,7 @@ void print_head (const char* str, size_t length)
 {
   const char* t;
 
-  for (t = str; t - str < length; putchar(*t), t++);
+  for (t = str; t - str < (int)length; putchar(*t), t++);
 }
 
 static
